label = value if empty
Hello, thanks for this module. I was wandering: wouldn't be better to show the value if the label is empty, otherwise the select element could have empty choices. This could be useful when the value and the label are not different or when you want to show the exact value in the label. kind regards.